+++ core/base/src/commonTest/kotlin/nl/adaptivity/xmlutil/TestKtXmlReader.kt
@@ -26,6 +26,7 @@ import nl.adaptivity.xmlutil.core.KtXmlWriter
 import nl.adaptivity.xmlutil.core.impl.multiplatform.StringReader
 import nl.adaptivity.xmlutil.core.impl.multiplatform.StringWriter
 import nl.adaptivity.xmlutil.core.impl.multiplatform.use
+import nl.adaptivity.xmlutil.test.TestCommonReader
 import kotlin.test.Test
 import kotlin.test.assertEquals